add errorHandler in utility func for ac
