+++ v0/destinations/active_campaign/transform.js
@@ -13,9 +13,7 @@ const {
   getErrorRespEvents,
   CustomError
 } = require("../../util");
-const {
-  nodeSysErrorToStatus
-} = require("../../../adapters/utils/networkUtils");
+const { errorHandler } = require("./util");
 
 // The Final data is both application/url-encoded FORM and POST JSON depending on type of event
 // Creating a switch case for final request building
@@ -87,18 +85,7 @@ const customTagProcessor = async (message, category, destination) => {
       }
     );
   } catch (err) {
-    if (err.response) {
-      throw new CustomError(
-        `Failed to create new contact (${err.response.statusText})`,
-        err.response.status || 400
-      );
-    } else {
-      const httpError = nodeSysErrorToStatus(err.code);
-      throw new CustomError(
-        `Failed to create new contact ${httpError.message}`,
-        httpError.status
-      );
-    }
+    errorHandler(err, "Failed to create new contact");
   }
   const createdContact = res.data.contact;
 
@@ -128,18 +115,7 @@ const customTagProcessor = async (message, category, destination) => {
       }
     );
   } catch (err) {
-    if (err.response) {
-      throw new CustomError(
-        `Failed to fetch already created tags (${err.response.statusText})`,
-        err.response.status || 400
-      );
-    } else {
-      const httpError = nodeSysErrorToStatus(err.code);
-      throw new CustomError(
-        `Failed to fetch already created tags ${httpError.message}`,
-        httpError.status
-      );
-    }
+    errorHandler(err, "Failed to fetch already created tags");
   }
 
   const storedTags = {};
@@ -181,18 +157,7 @@ const customTagProcessor = async (message, category, destination) => {
           );
           promises.push(resp);
         } catch (err) {
-          if (err.response) {
-            throw new CustomError(
-              `Failed to fetch already created tags (${err.response.statusText})`,
-              err.response.status || 400
-            );
-          } else {
-            const httpError = nodeSysErrorToStatus(err.code);
-            throw new CustomError(
-              `Failed to fetch already created tags ${httpError.message}`,
-              httpError.status
-            );
-          }
+          errorHandler(err, "Failed to fetch already created tags");
         }
       }
       const results = await Promise.all(promises);
@@ -241,18 +206,7 @@ const customTagProcessor = async (message, category, destination) => {
             }
           );
         } catch (err) {
-          if (err.response) {
-            throw new CustomError(
-              `Failed to create tags (${err.response.statusText})`,
-              err.response.status || 400
-            );
-          } else {
-            const httpError = nodeSysErrorToStatus(err.code);
-            throw new CustomError(
-              `Failed to create tags ${httpError.message}`,
-              httpError.status
-            );
-          }
+          errorHandler(err, "Failed to create tags");
         }
         // For each tags successfully created the response id is pushed to tagIds
         if (res.status === 201) tagIds.push(res.data.tag.id);
@@ -286,18 +240,7 @@ const customTagProcessor = async (message, category, destination) => {
           }
         );
       } catch (err) {
-        if (err.response) {
-          throw new CustomError(
-            `Failed to merge created contact with created tags (${err.response.statusText})`,
-            err.response.status || 400
-          );
-        } else {
-          const httpError = nodeSysErrorToStatus(err.code);
-          throw new CustomError(
-            `Failed to merge created contact with created tags ${httpError.message}`,
-            httpError.status
-          );
-        }
+        errorHandler(err, "Failed to merge created contact with created tags");
       }
     })
   );
@@ -340,18 +283,7 @@ const customFieldProcessor = async (
     );
     responseStaging = res.status === 200 ? res.data.fields : [];
   } catch (err) {
-    if (err.response) {
-      throw new CustomError(
-        `Failed to get existing field data (${err.response.statusText})`,
-        err.response.status || 400
-      );
-    } else {
-      const httpError = nodeSysErrorToStatus(err.code);
-      throw new CustomError(
-        `Failed to get existing field data ${httpError.message}`,
-        httpError.status
-      );
-    }
+    errorHandler(err, "Failed to get existing field data");
   }
 
   // From the responseStaging we store the stored field information in K-V struct iin fieldMap
@@ -410,18 +342,7 @@ const customFieldProcessor = async (
           }
         );
       } catch (err) {
-        if (err.response) {
-          throw new CustomError(
-            `Failed to create mapping request (${err.response.statusText})`,
-            err.response.status || 400
-          );
-        } else {
-          const httpError = nodeSysErrorToStatus(err.code);
-          throw new CustomError(
-            `Failed to create mapping request ${httpError.message}`,
-            httpError.status
-          );
-        }
+        errorHandler(err, "Failed to create mapping request");
       }
     })
   );
@@ -477,18 +398,7 @@ const customListProcessor = async (
             }
           );
         } catch (err) {
-          if (err.response) {
-            throw new CustomError(
-              `Failed to map created contact with the list (${err.response.statusText})`,
-              err.response.status || 400
-            );
-          } else {
-            const httpError = nodeSysErrorToStatus(err.code);
-            throw new CustomError(
-              `Failed to map created contact with the list ${httpError.message}`,
-              httpError.status
-            );
-          }
+          errorHandler(err, "Failed to map created contact with the list");
         }
       } else {
       }
@@ -543,18 +453,7 @@ const screenRequestHandler = async (message, category, destination) => {
       }
     );
   } catch (err) {
-    if (err.response) {
-      throw new CustomError(
-        `Failed to retrieve events (${err.response.statusText})`,
-        err.response.status || 400
-      );
-    } else {
-      const httpError = nodeSysErrorToStatus(err.code);
-      throw new CustomError(
-        `Failed to retrieve events ${httpError.message}`,
-        httpError.status
-      );
-    }
+    errorHandler(err, "Failed to retrieve events");
   }
 
   if (res.status !== 200)
@@ -587,18 +486,7 @@ const screenRequestHandler = async (message, category, destination) => {
         }
       );
     } catch (err) {
-      if (err.response) {
-        throw new CustomError(
-          `Failed to create event (${err.response.statusText})`,
-          err.response.status || 400
-        );
-      } else {
-        const httpError = nodeSysErrorToStatus(err.code);
-        throw new CustomError(
-          `Failed to create event ${httpError.message}`,
-          httpError.status
-        );
-      }
+      errorHandler(err, "Failed to create event");
     }
 
     if (res.status !== 201) {
@@ -613,11 +501,8 @@ const screenRequestHandler = async (message, category, destination) => {
   payload.actid = destination.Config.actid;
   payload.key = destination.Config.eventKey;
   payload.visit = encodeURIComponent(
-    `{"email" : "${
-      message.context.traits.email
-        ? message.context.traits.email
-        : message.context.traits.traits.email
-    }"}`
+    `{"email":"${get(message, "context.traits.email") ||
+      get(message, "context.traits.traits.email")}"}`
   );
   return responseBuilderSimple(payload, category, destination);
 };
@@ -639,18 +524,7 @@ const trackRequestHandler = async (message, category, destination) => {
       }
     );
   } catch (err) {
-    if (err.response) {
-      throw new CustomError(
-        `Failed to retrieve events (${err.response.statusText})`,
-        err.response.status || 400
-      );
-    } else {
-      const httpError = nodeSysErrorToStatus(err.code);
-      throw new CustomError(
-        `Failed to retrieve events ${httpError.message}`,
-        httpError.status
-      );
-    }
+    errorHandler(err, "Failed to retrieve events");
   }
 
   if (res.status !== 200)
@@ -683,18 +557,7 @@ const trackRequestHandler = async (message, category, destination) => {
         }
       );
     } catch (err) {
-      if (err.response) {
-        throw new CustomError(
-          `Failed to create event (${err.response.statusText})`,
-          err.response.status || 400
-        );
-      } else {
-        const httpError = nodeSysErrorToStatus(err.code);
-        throw new CustomError(
-          `Failed to create event ${httpError.message}`,
-          httpError.status
-        );
-      }
+      errorHandler(err, "Failed to create event");
     }
 
     if (res.status !== 201) {
@@ -709,11 +572,8 @@ const trackRequestHandler = async (message, category, destination) => {
   payload.actid = destination.Config.actid;
   payload.key = destination.Config.eventKey;
   payload.visit = encodeURIComponent(
-    `{"email" : "${
-      message.context.traits.email
-        ? message.context.traits.email
-        : message.context.traits.traits.email
-    }"}`
+    `{"email":"${get(message, "context.traits.email") ||
+      get(message, "context.traits.traits.email")}"}`
   );
   return responseBuilderSimple(payload, category, destination);
 };
